power: supply: qcom_fg: remove PMI8994 fuel gauge
Remove Pre-Gen3 fuel gauge support the two generation have very
little in common. This driver squashing 2 driver together so
better to separate them.

Additional Changes:
- Fix all checkpatch issues in strict mode.
- Rename the driver to pmi8998_fg.
- Remove unneccessary batt_info.
