+++ drivers/power/supply/Kconfig
@@ -720,6 +720,14 @@ config CHARGER_PM8916_LBC
 	  To compile this driver as module, choose M here: the
 	  module will be called pm8916_lbc.
 
+config BATTERY_PMI8998_FG
+	tristate "Qualcomm PMI8998 PMIC fuel gauge driver"
+	depends on MFD_SPMI_PMIC
+	help
+	  Say Y here to enable the Qualcomm PMI8998 PMIC Fuel Gauge driver.
+	  This adds support for battery fuel gauging and state of charge of
+	  battery connected to the fuel gauge.
+
 config CHARGER_BQ2415X
 	tristate "TI BQ2415x battery charger driver"
 	depends on I2C
@@ -1054,14 +1062,6 @@ config BATTERY_UG3105
 	  Since Linux does not monitor coulomb-counter changes while the
 	  device is off or suspended, the functionality of this driver is
 	  limited to reporting capacity only.
-
-config BATTERY_QCOM_FG
-	tristate "Qualcomm PMIC fuel gauge driver"
-	depends on MFD_SPMI_PMIC
-	help
-	  Say Y here to enable the Qualcomm PMIC Fuel Gauge driver. This
-	  adds support for battery fuel gauging and state of charge of
-	  battery connected to the fuel gauge.
 
 config CHARGER_QCOM_SMB2
 	tristate "Qualcomm PMI8998 PMIC charger driver"
